fix(ci): python-coverage-shards — extract guide_workspace so controls.py stays under the 1500-line maintainability limit
The #447 guide control pushed controls.py to 1617 lines, tripping
test_first_party_code_files_stay_under_line_limit. Move guide_workspace
into a focused controls_guide.py module wired back via _WorkspaceGuideMixin,
mirroring the executor/worker delegate-mixin decomposition. Behavior unchanged.
